(24.06.07 11:29) 에러 처리, api 명세서 양식으로 작성.

+++ src/main/java/com/green/gittest/common/CheckMapper.java
@@ -7,6 +7,6 @@ public interface CheckMapper {
     Long getUserId(long userId);
     Long getPetId(long petId);
     Long getCalendarId(long calendarId);
-    int getListId(long listId);
-    int getToDoListIsCompleted(long listId);
+    Long getListId(long listId);
+    Integer getToDoListIsCompleted(long listId);
 }
